backwards-compatible with old mapbox rendering code
