+++ pixiedust/display/chart/__init__.py
@@ -23,7 +23,7 @@ import pixiedust
 myLogger = pixiedust.getLogger(__name__ )
 
 #bootstrap all the renderers
-renderers = ["matplotlib", "bokeh", "seaborn", "mapbox-points", "mapbox", "google", "brunel", "table"]
+renderers = ["matplotlib", "bokeh", "seaborn", "mapbox", "google", "brunel", "table"]
 
 for renderer in renderers:
     try:
