+++ pixiedust/display/chart/renderers/mapbox-old/mapBoxBaseDisplay.py
@@ -23,14 +23,14 @@ import pixiedust
 
 myLogger = pixiedust.getLogger(__name__)
 
-@PixiedustRenderer(rendererId="mapbox-points")
+@PixiedustRenderer(rendererId="mapbox")
 class MapBoxBaseDisplay(with_metaclass(ABCMeta, BaseChartDisplay)):
     def get_options_dialog_pixieapp(self):
             """
             Return the fully qualified path to a PixieApp used to display the dialog options
             PixieApp must inherit from pixiedust.display.chart.options.baseOptions.BaseOptions
             """
-            return "pixiedust.display.chart.renderers.mapbox-points.mapboxOptions.MapboxOptions"
+            return "pixiedust.display.chart.renderers.mapbox.mapboxOptions.MapboxOptions"
 
     def getChartOptions(self):
         return [ 
@@ -59,7 +59,7 @@ class MapBoxBaseDisplay(with_metaclass(ABCMeta, BaseChartDisplay)):
                         'type': 'slider',
                         'max': 100,
                         'min': 0,
-                        'default': 50
+                        'default': 80
                 }
             },
             {
