+++ pixiedust/display/chart/renderers/mapbox-old/templates/mapViewOptionsDialogBody.html
